Add ServiceWorker for offline assets
This adds broccoli-serviceworker, which automatically generates a
ServiceWorker file and registration for the Ember app assets
(cache-first) and the Sockethub assets (network-first). Configuration
is in `config/environment.js`.

We'll need our own SW for Push notifications and more sophisticated cash
handling (e.g. for Sockethub assets), of course, so this is meant mostly
for loading time improvements in production for now.