WIP Mark messages as grouped when appropriate
When a message is sent/received from the same person quickly after the
last one, and with no message in between, this marks them as grouped.

The main purpose is to visually group the message under the same name
and avatar, instead of adding all metadata to every line of chat.